Make ConcertActorClusterSharding top level class
